moved store folder to src folder and updated imports
