+++ src/store/bitcoinCharts/index.js
