+++ src/store/coinPage/actions.js
